% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_mm.R
\name{beta_mm}
\alias{beta_mm}
\title{Obtain the shape parameters of a beta distribution}
\usage{
beta_mm(p)
}
\arguments{
\item{p}{a vector of iid beta random variables, for which you want
to obtain the parameters of the beta distribution from which they
are drawn}
}
\description{
Use the method of moments to obtain the \code{shape1} and \code{shape2}
parameters when provided a vector assumed iid beta random variables.
}
