% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize the first letter in a string}
\usage{
capitalize(x)
}
\arguments{
\item{x}{a character vector of any length}
}
\description{
Like \code{toupper}, but only for the first letter of a string.
}
\examples{
capitalize("ben")

}
