% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv2sig.R
\name{cv2sig}
\alias{cv2sig}
\title{Convert CV to Lognormal SD}
\usage{
cv2sig(cv)
}
\arguments{
\item{cv}{a numeric vector, possibly length > 1,
containing the CV to convert.}
}
\value{
a numeric vector with the same length as \code{cv},
  the lognormal standard deviation.
}
\description{
Used to convert the coefficient of variation of a
random variable to the lognormal standard deviation,
using a simple transformation.
}
\examples{
cv2sig(cv = 0.2)

}
\seealso{
\code{\link{sig2cv}}
}
