% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_time2nice_time.R
\name{dec_time2nice_time}
\alias{dec_time2nice_time}
\title{Convert a decimal time to a nice-looking time}
\usage{
dec_time2nice_time(dec_times, clock = "12hr")
}
\arguments{
\item{dec_times}{a numeric vector with length >= 1,
and values in the interval [0,24].}

\item{clock}{a character vector specifying if you want a 12-hour clock
(\code{"12hr"}) or a 24-hour clock(i.e., miliary time; \code{"24hr"})}
}
\value{
a character vector with the converted times.
  If \code{clock == "12hr"}, either \code{"AM"} or \code{"PM"}
  will be included at the end of the character string in each element.
}
\description{
Goes from hours.fraction to hours:minutes.
}
\note{
Any cases where \code{dec_times == 0 | dec_times == 24} will be
  converted to \code{"12:00AM"} if \code{clock == "12hr"} and
  \code{"0:00"} if \code{clock == "24hr"}.
}
\examples{
dec_time2nice_time(c(2.25, 14.55), "12hr")
dec_time2nice_time(c(2.25, 14.55), "24hr")
}
