% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit.R
\name{expit}
\alias{expit}
\title{Perform the Inverse-Logit Tranformation}
\usage{
expit(x)
}
\arguments{
\item{x}{A numeric vector with length >= 1 containing the number
in logit space to convert to the [0,1] scale.}
}
\value{
A numeric vector with the same length as \code{x} containing the inverse-logit of \code{x}.
}
\description{
This function performs the inverse-logit transformation.
See the details below for exactly what this means.
}
\details{
The inverse logit tranformation turns any real number (negative to positive infinity) to
  a number on the [0,1] scale (e.g., the probability scale).
}
\examples{
expit(x = 1)
expit(x = 0.5)
expit(x = logit(x = 0.5))

}
\seealso{
\code{\link{logit}}
}
