% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_device.R
\name{ext_device}
\alias{ext_device}
\title{Create a new external plotting device}
\usage{
ext_device(h = 7, w = 7, title = NULL)
}
\arguments{
\item{h}{the height, in inches}

\item{w}{the width, in inches}

\item{title}{the desired title of the plotting window}
}
\description{
This is a platform-flexible function for creating a new plotting device.
It functions just like \code{windows()} on Windows OS, \code{quartz()} on Mac OS,
and \code{x11()} on Linux OS.
}
