% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Opposite of \code{\%in\%}}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\description{
Sometimes is nice to have a shortcut.
}
\details{
the infix operator \code{\%in\%} tests whether elements in one vector
  are also found in another vector. To negate this test, previously needed to wrap
  the whole statement in parentheses. This helps you get around this hassle.
}
\examples{
v = c("a", "b", "c"); "x" \%!in\% v

}
