% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_leap_year.R
\name{is_leap_year}
\alias{is_leap_year}
\title{Determine if a year is a leap year}
\usage{
is_leap_year(year)
}
\arguments{
\item{date}{a numeric vector of length >= 1 containing the years to test}
}
\value{
a logical vector indicating whether each year is a leap year (\code{TRUE} if so, \code{FALSE} if not).
}
\description{
Used to determine if each element in a vector of years is a leap year.
}
\details{
a year is considered a leap year if it is divisible by 4 and not divisible by 100.
  The exception is years that are divisible by 400 are also considered leap years.
}
\examples{
is_leap_year(2016:2020)

}
