% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify.R
\name{prettify}
\alias{prettify}
\title{Add commas to a numeric vector}
\usage{
prettify(x)
}
\arguments{
\item{x}{a numeric vector to be prettified.}
}
\value{
a character vector with the prettified output.
}
\description{
This function is a wrapper for \code{prettyNum()}, and is used only to
add a comma to separate thousands places.
}
\examples{
prettify(10000)

}
\seealso{
\code{\link[base]{prettyNum}}, \code{\link{summ}}
}
