% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_updater.R
\name{progress_updater}
\alias{progress_updater}
\title{Print a progress indicator to console}
\usage{
progress_updater(i, n, grp = NULL, rnd = 0)
}
\arguments{
\item{i}{numeric vector of length == 1. The current iteration value}

\item{n}{numeric vector of length == 1. The number of iterations to calculate}

\item{grp}{character or numeric vector of length == 1. Optional additional identifier to be placed
in parentheses following the printed percent. Defaults to \code{NULL}, in which case no additional
group identifier is printed}

\item{rnd}{numeric vector of length == 1. Controls the number of decimal values in the printed percent.
Defaults to \code{rnd = 0}, in which case only the integer value percents will be printed.
Set to \code{rnd = 1} to obtain tenths of a percent, \code{rnd = 2} to obtain hundredths, etc.}
}
\description{
Include this function in a loop or other calculation that iterates over elements
in which the total number of elements is known to obtain an indication of what percentage
of the calculation is completed
}
\examples{
 junk = sapply(1:15, function(i) {progress_updater(i, 15, rnd = 1); Sys.sleep(0.2)})
}
