% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig2cv.R
\name{sig2cv}
\alias{sig2cv}
\title{Convert Lognormal SD to CV}
\usage{
sig2cv(sig)
}
\arguments{
\item{cv}{A numeric vector, possibly length > 1, containing the lognormal SD you wish to convert.}
}
\value{
A numeric vector with the same length as \code{sig},
  the coefficient of variation.
}
\description{
Used to convert the lognormal standard deviation
of a random variable to the coefficient of variation
using a simple transformation.
}
\examples{
sig2cv(sig = 0.2)

}
\seealso{
\code{\link{cv2sig}}
}
