% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatonMisc.R
\docType{package}
\name{StatonMisc}
\alias{StatonMisc}
\alias{StatonMisc-package}
\title{StatonMisc: Common Functions Used by B.A. Staton}
\description{
This package contains a somewhat random assortment of functions that are
commonly used in code written by Ben Staton, including his other packages and
non-public code. It includes mostly utility-type functions for tasks like
transforming dates (from MM/DD/YYYY to DOY and back) and basic data
transformations (\code{cv2sig}, \code{sig2cv}, \code{logis}, \code{expit}, etc.).
}
