% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_in_dir.R
\name{string_in_dir}
\alias{string_in_dir}
\title{Find string occurrences in files within a directory}
\usage{
string_in_dir(path = NULL, string, print_negs = F, print_lines = F, ...)
}
\arguments{
\item{path}{a character string (length == 1) pointing to the location of the directory to search.
Passed to \code{dir(path)}. Defaults to \code{NULL}, which specifies the working directory}

\item{string}{a character string (length == 1) indicating the string to search for
Regular expressions are supported here}

\item{print_negs}{logical. Do you wish to see the names of the files were searched but
that do not have any matches? Defaults to \code{FALSE}}

\item{print_lines}{logical or \code{"ask"}. Do you wish to see the specific line matches in each file?
Specifying \code{"ask"} will ask the user before printing. Defaults to \code{FALSE}}

\item{...}{optional arguments to be passed to \code{dir()} when selecting files to search.
For example, \code{recursive = TRUE} will search files within subdirectories,
and \code{pattern = ".R|.Rmd"} will search only within files with these extensions.}
}
\value{
The search results are printed to the console.
}
\description{
Finds the location, frequency, and context with which a given string is used
in files contained within a specified directory. Regular expressions are
supported to facilitate searching in both string location and file selection.
}
\seealso{
\code{\link[base]{dir}}
}
