% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_mm.R
\name{beta_mm}
\alias{beta_mm}
\title{Obtain the shape parameters of a beta distribution}
\usage{
beta_mm(mn = NULL, vr = NULL, p = NULL, na.rm = FALSE)
}
\arguments{
\item{mn}{numeric vector of length == 1. The mean of a beta random variable.
Must be specified if \code{vr} is not \code{NULL}}

\item{vr}{numeric vector of length == 1. The variance of a beta random variable.
Must be specified if \code{mn} is not \code{NULL}}

\item{p}{numeric vector of length > 1. iid beta random variables.
Must be \code{NULL} if \code{mn} and \code{vr} are not \code{NULL}}

\item{na.rm}{logical. Do you wish to remove NA elements before calculating mean and variance?
Defaults to \code{FALSE} and only takes effect if \code{p} is not \code{NULL}.}
}
\description{
Use the method of moments to obtain the \code{shape1} (alpha) and \code{shape2} (beta)
parameters when provided either a vector of beta random variables
or the sample mean and variance of such a vector.
}
\examples{
beta_mm(p = rbeta(1e6, 6, 1))
beta_mm(mn = 0.5, vr = 0.04)
}
