% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date2doy.R
\name{date2doy}
\alias{date2doy}
\title{Convert Date to DOY}
\usage{
date2doy(date)
}
\arguments{
\item{date}{a character vector containing the date to convert. June 1 2018 would be \code{"6/1/2018"},
and December 25 1975 would be \code{"12/25/1975"}. Currently, this function
can only use dates from a single year.}
}
\value{
a numeric vector containing the DOY that was converted.
}
\description{
This function is used to convert a vector of dates formated as M-D-YYYY
to a vector of days of the year. The calculation automatically takes
leap years into account.
}
\examples{
date2doy("6/12/2016")
date2doy(c("6/12/2017", "6/19/2017"))

}
\seealso{
\code{\link{doy2date}}
}
