% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doy2date.R
\name{doy2date}
\alias{doy2date}
\title{Convert DOYs to Dates}
\usage{
doy2date(doy, year, leading_zero = FALSE, include_year = TRUE)
}
\arguments{
\item{doy}{A numeric vector of length 1 specifying the day of the year to convert.}

\item{year}{A numeric vector of length 1 specifying the year \code{doy} corresponds to.
This is necessary to account for leap years.}

\item{leading_zero}{logical: Do you wish to insert a zero in front of the month and day in the date
if they are just single numbers (e.g., \code{"6/1"} would become \code{"06/01"} if set to \code{TRUE}).
Defaults to \code{FALSE}.}

\item{include_year}{logical: Do you wish to include the year in the output?
Defaults to \code{TRUE}.}
}
\value{
A character vector containing the converted dates.
}
\description{
This function is used to convert a vector of dates formated as M-D-YYYY
to a vector of days of the year, taking leap years into account.
}
\examples{
doy2date(51, 2016)
date2doy(51, 2017, TRUE)

}
\seealso{
\code{\link{date2doy}}
}
