% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_device.R
\name{file_device}
\alias{file_device}
\title{Create a graphics device file based on file extension}
\usage{
file_device(file, width = 5, height = 5, res = 600)
}
\arguments{
\item{file}{the name of the output file. Can be a path,
but must contain a file extension at the end (e.g., \code{"path/to/dir/file.png"})}

\item{width}{the width of the device, in inches. Defaults to 5}

\item{height}{the height of the device, in inches. Defaults to 5}

\item{res}{the resolution of the device, in pixels per inch. Defaults to 600}
}
\description{
Wouldn't it be nice to switch between (e.g.,) pdf and jpg
files by simply changing the file extension?
}
\details{
Accepted file extentions include:
\itemize{
  \item \code{".pdf"}
  \item \code{".png"}
  \item \code{".bmp"}
  \item \code{".jpg"} or \code{".jpeg"}
  \item \code{".tif"} or \code{".tiff"}
}
}
\examples{
\dontrun{
# use a png device
file_device("test.png")
hist(rnorm(1000), col = "grey")
dev.off()

# use a pdf device: can dump multiple plots in one file!
file_device("test.pdf")
hist(rnorm(1000), col = "grey")
boxplot(rnorm(1000), col = "grey")
dev.off()
}
}
\seealso{
\code{\link[grDevices]{pdf}}, \code{\link[grDevices]{jpeg}},
  \code{\link[grDevices]{png}}, \code{\link[grDevices]{tiff}}, \code{\link[grDevices]{bmp}}
}
