% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_leap_year.R
\name{is_leap_year}
\alias{is_leap_year}
\title{Determine if a year is a leap year}
\usage{
is_leap_year(year)
}
\arguments{
\item{date}{a numeric vector of length >= 1 containing the years to test}
}
\value{
a logical vector indicating whether each year is a leap year (TRUE if so, FALSE if not).
}
\description{
This function is used to determine if each element in a vector of years is
a leap year.
}
\examples{
is_leap_year(2016:2020)
}
