% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ.R
\name{summ}
\alias{summ}
\title{Summarize a Vector of Random Quantities}
\usage{
summ(x, p = c(0.5, 0.025, 0.975), rnd = NULL, na.rm = F,
  prettify = FALSE)
}
\arguments{
\item{x}{a numeric vector with length > 1. It can have \code{NA} elements.}

\item{p}{numeric vector with elements on the interval (0,1).
The quantile points requested. Defaults to \code{c(0.5, 0.025, 0.975)}}

\item{rnd}{a numeric vector of length == 1. The rounding specification.
Defaults to \code{NULL}, i.e., no rounding.}

\item{na.rm}{logical. Do you wish to remove NAs before calculating summary statistics?
Defaults to \code{FALSE}.}

\item{prettify}{logical. Do you wish to insert commas in the output? If so,
it will be coerced to a character vector. Defaults to \code{FALSE}.}
}
\value{
a vector with length depending on the requested quantiles.
  If \code{isTRUE(prettify)}, output will be a named character vector, otherwise,
  it will be a named numeric vector.
}
\description{
This function is a wrapper for \code{mean()}, \code{quantile()},
\code{sd}, \code{round()}, and \code{prettyNum()}. So it can easily
be called with less code. Also, nice to be able place these calls in
\code{apply()} call.
}
\examples{
summ(x = rnorm(100, 50000, 5000))
summ(x = rnorm(100, 50000, 5000),
     p = c(0.025, 0.1, 0.25, 0.5, 0.75, 0.9, 0.975),
     rnd = -3, prettify = T)
}
\seealso{
\code{\link[base]{round}}, \code{\link[base]{quantile}}
}
